smoketest: T5160: Deduplicate nftables verify functions to testcase class, remove obsolete imports
